Rename TestUtils.h ti DemosUtils.h
