+++ demos/cpp/demo-kineticpath-perturbed-co2.cpp
@@ -20,7 +20,7 @@
 using namespace Reaktoro;
 
 // Reactive transport test includes
-#include <demos/cpp/TestUtils.h>
+#include <demos/cpp/DemosUtils.h>
 
 auto runKinetics(KineticPathParams &) -> void;
